requirements.txt
jupyter
pandas
numpy
matplotlib
scipy
yfinance
binance-connector